media-source/MediaSource-addSourceBuffer_null.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>MediaSource#addSourceBuffer(null) and MediaSource#addSourceBuffer("")</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>

<script>
async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e) {
        var sourceBuffer = null;
        t.step(function() {
            assert_throws(new TypeError(),
                          function() {
                              sourceBuffer = mediaSource.addSourceBuffer(null);
                          },
                          'Set null');
        });
        t.step(function() {
            assert_throws({name : 'InvalidAccessError'},
                          function() {
                              sourceBuffer = mediaSource.addSourceBuffer('');
                          },
                          'Set a empty string');
        });
        t.done();
        mediaSource.endOfStream();
    }, false);
    var video = document.createElement('video');
    video.src = window.URL.createObjectURL(mediaSource);
}, 'MediaSource#addSourceBuffer : If type is null or an empty string then throw an INVALID_ACCESS_ERR exception and abort these steps.');
</script>
</body>
</html>
